More fixes

+++ bootstrap/vm-kind.sh
@@ -16,12 +16,20 @@ function bootstrap_precheck() {
     fi
 
     # TODO: Check for existence
-    hcloud ssh-key create --name "${CLUSTER_NAME}-jump" --public-key-from-file ./ssh-jump.pub --label type="${CLUSTER_NAME}-jump"
+    hcloud ssh-key create \
+        --name="${CLUSTER_NAME}-jump" \
+        --public-key-from-file=./ssh-jump.pub \
+        --label=type="${CLUSTER_NAME}-jump"
     # TODO: Check for existence
-    hcloud server create --name "${CLUSTER_NAME}-jump" --type cx32 --image ubuntu-24.04 --ssh-key "${CLUSTER_NAME}-jump" --label type="${CLUSTER_NAME}-jump"
+    hcloud server create \
+        --name="${CLUSTER_NAME}-jump" \
+        --type=cx32 \
+        --image=ubuntu-24.04 \
+        --ssh-key="${CLUSTER_NAME}-jump" \
+        --label=type="${CLUSTER_NAME}-jump"
 
     JUMP_IP="$(
-        hcloud server list --selector type="${CLUSTER_NAME}-jump" --output json \
+        hcloud server list --selector=type="${CLUSTER_NAME}-jump" --output=json \
         | jq -r '.[].public_net.ipv4.ip'
     )"
     cat >"${HOME}/.ssh/config.d/${CLUSTER_NAME}-jump" <<-EOF
@@ -83,9 +91,9 @@ function bootstrap_create() {
         cat bootstrap/kind.yaml \
         | ssh "${CLUSTER_NAME}-jump" \
             kind create cluster \
-                --name "${name}" \
-                --config - \
-                --wait 5m
+                --name="${name}" \
+                --config=- \
+                --wait=5m
     fi
 }
 
@@ -102,7 +110,25 @@ function bootstrap_delete() {
         hcloud ssh-key delete "${CLUSTER_NAME}-jump"
     fi
 
-    kill "${VM_KIND_TUNNEL_PID}"
+    if test -z "${VM_KIND_TUNNEL_PID}"; then
+        PORT="$(
+            kubectl --kubeconfig=./kubeconfig-bootstrap config view --output=json \
+            | jq --raw-output --arg name "${name}" '.clusters[] | select(.name == "kind-\($name)") | .cluster.server' \
+            | cut -d: -f3
+        )"
+        VM_KIND_TUNNEL_PID="$(
+            netstat -tunapl 2>/dev/null \
+            | grep LISTEN \
+            | grep "127.0.0.1:${POST}" \
+            | tr -s " " \
+            | cut -d' ' -f7 \
+            | cut -d/ -f1
+        )"
+    fi
+    if test -n "${VM_KIND_TUNNEL_PID}"; then
+        echo "### Killing tunnel process"
+        kill "${VM_KIND_TUNNEL_PID}"
+    fi
 }
 
 function bootstrap_kubeconfig() {
